Refactor to be more generic
