--- config/element_filters.yaml
@@ -1,15 +1,0 @@
-services:
-  _defaults:
-    autowire: true
-    autoconfigure: true
-    public: false
-
-  #Filter
-
-  #Common
-  Pimcore\Bundle\StudioBackendBundle\Element\Filter\FilterLoaderInterface:
-    class:  Pimcore\Bundle\StudioBackendBundle\Element\Filter\Loader\TaggedIteratorAdapter
-
-  Pimcore\Bundle\StudioBackendBundle\Element\Service\ElementListingFilterInterface:
-    class: Pimcore\Bundle\StudioBackendBundle\Element\Service\ElementListingFilter
-    tags: [ 'pimcore.studio_backend.filter_service' ]
