+++ src/Grid/MappedParameter/GridParameter.php
@@ -16,6 +16,8 @@ declare(strict_types=1);
 
 namespace Pimcore\Bundle\StudioBackendBundle\Grid\MappedParameter;
 
+use Pimcore\Bundle\StudioBackendBundle\Filter\MappedParameter\FilterParameter;
+
 /**
  * @internal
  */
