+++ src/Listing/Filter/Filters.php
@@ -14,18 +14,20 @@ declare(strict_types=1);
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Bundle\StudioBackendBundle\Element\Filter;
-
-use Pimcore\Model\Listing\AbstractListing;
-use Pimcore\Model\Listing\CallableFilterListingInterface;
+namespace Pimcore\Bundle\StudioBackendBundle\Listing\Filter;
 
 /**
  * @internal
  */
-interface FilterInterface
+final readonly class Filters
 {
-    public function apply(
-        mixed $parameters,
-        AbstractListing|CallableFilterListingInterface $listing
-    ): AbstractListing|CallableFilterListingInterface;
+    public function __construct(
+        private array $filters = [],
+    ) {
+    }
+
+    public function getFilters(): array
+    {
+        return $this->filters;
+    }
 }
